# api: improve performance of ical endpoint