Using cProfile, the icalendar generation portion is now around 8-9ms instead of 80ms

Before:

```
❯ ./.venv/bin/python -m calgen                                                         0.18s
         292168 function calls (291774 primitive calls) in 0.080 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.001    0.001    0.080    0.080 icalendar_utils.py:34(create_calendar)
        1    0.000    0.000    0.063    0.063 cal.py:420(to_ical)
        1    0.002    0.002    0.051    0.051 cal.py:410(content_lines)
     4675    0.003    0.000    0.038    0.000 cal.py:404(content_line)
     4675    0.007    0.000    0.032    0.000 parser.py:300(from_parts)
    10520    0.007    0.000    0.014    0.000 caselessdict.py:30(__init__)
      389    0.001    0.000    0.012    0.000 icalendar_utils.py:9(_create_event)
      779    0.000    0.000    0.011    0.000 {method 'join' of 'bytes' objects}
        1    0.000    0.000    0.011    0.011 parser.py:377(to_ical)
     4676    0.001    0.000    0.011    0.000 parser.py:380(<genexpr>)
    390/1    0.003    0.000    0.011    0.011 cal.py:293(property_items)
    31553    0.008    0.000    0.010    0.000 parser_tools.py:9(to_unicode)
     4675    0.002    0.000    0.010    0.000 parser.py:365(to_ical)
     4675    0.004    0.000    0.008    0.000 parser.py:65(foldline)
      390    0.000    0.000    0.007    0.000 cal.py:79(__init__)
     3510    0.001    0.000    0.007    0.000 prop.py:726(to_ical)
     3510    0.003    0.000    0.006    0.000 prop.py:716(__new__)
     3510    0.003    0.000    0.005    0.000 parser.py:20(escape_char)
     4285    0.003    0.000    0.005    0.000 caselessdict.py:40(__getitem__)
     4673    0.003    0.000    0.005    0.000 caselessdict.py:44(__setitem__)
     4675    0.003    0.000    0.004    0.000 parser.py:292(__new__)
      389    0.000    0.000    0.003    0.000 functional.py:224(wrapper)
    58070    0.003    0.000    0.003    0.000 {built-in method builtins.isinstance}
      778    0.000    0.000    0.003    0.000 prop.py:355(__init__)
      389    0.001    0.000    0.003    0.000 text.py:397(slugify)
     4675    0.001    0.000    0.003    0.000 {method 'join' of 'str' objects}
      778    0.001    0.000    0.003    0.000 parser.py:215(to_ical)
     1167    0.001    0.000    0.002    0.000 prop.py:393(to_ical)
      778    0.000    0.000    0.002    0.000 __init__.py:178(sub)
    21838    0.002    0.000    0.002    0.000 {method 'replace' of 'str' objects}
    17422    0.001    0.000    0.001    0.000 {method 'encode' of 'str' objects}
     9393    0.001    0.000    0.001    0.000 parser.py:83(<genexpr>)
     8185    0.001    0.000    0.001    0.000 {built-in method __new__ of type object at 0x1005a7308}
     1167    0.000    0.000    0.001    0.000 prop.py:389(__init__)
     1167    0.001    0.000    0.001    0.000 parser.py:52(tzid_from_dt)
    14018    0.001    0.000    0.001    0.000 {method 'upper' of 'str' objects}
      778    0.000    0.000    0.001    0.000 parser.py:103(param_value)
      778    0.001    0.000    0.001    0.000 {method 'sub' of 're.Pattern' objects}
    11298    0.001    0.000    0.001    0.000 {method 'items' of 'collections.OrderedDict' objects}
    10864    0.001    0.000    0.001    0.000 {method 'append' of 'list' objects}
      778    0.000    0.000    0.001    0.000 __init__.py:272(_compile)
      778    0.001    0.000    0.001    0.000 prop.py:361(to_ical)
     6622    0.001    0.000    0.001    0.000 {method 'decode' of 'bytes' objects}
     7009    0.001    0.000    0.001    0.000 {built-in method builtins.hasattr}
      778    0.000    0.000    0.000    0.000 parser.py:142(dquote)
     4675    0.000    0.000    0.000    0.000 {built-in method builtins.getattr}
     4285    0.000    0.000    0.000    0.000 {function CaselessDict.__getitem__ at 0x10141ad40}
      389    0.000    0.000    0.000    0.000 {built-in method builtins.any}
5368/5365    0.000    0.000    0.000    0.000 {built-in method builtins.len}
        2    0.000    0.000    0.000    0.000 _compiler.py:738(compile)
      778    0.000    0.000    0.000    0.000 functional.py:226(<genexpr>)
      389    0.000    0.000    0.000    0.000 cal.py:267(add_component)
     1167    0.000    0.000    0.000    0.000 {method 'tzname' of 'datetime.timezone' objects}
      778    0.000    0.000    0.000    0.000 {built-in method fromisoformat}
      778    0.000    0.000    0.000    0.000 {method 'search' of 're.Pattern' objects}
        2    0.000    0.000    0.000    0.000 _parser.py:970(parse)
      389    0.000    0.000    0.000    0.000 {built-in method unicodedata.normalize}
        2    0.000    0.000    0.000    0.000 _compiler.py:571(_code)
        2    0.000    0.000    0.000    0.000 _parser.py:447(_parse_sub)
      389    0.000    0.000    0.000    0.000 {method 'strip' of 'str' objects}
        2    0.000    0.000    0.000    0.000 _parser.py:507(_parse)
        2    0.000    0.000    0.000    0.000 _compiler.py:509(_compile_info)
      389    0.000    0.000    0.000    0.000 {method 'values' of 'dict' objects}
      389    0.000    0.000    0.000    0.000 {method 'lower' of 'str' objects}
      390    0.000    0.000    0.000    0.000 {method 'keys' of 'collections.OrderedDict' objects}
      3/2    0.000    0.000    0.000    0.000 _compiler.py:37(_compile)
        3    0.000    0.000    0.000    0.000 _compiler.py:241(_optimize_charset)
        4    0.000    0.000    0.000    0.000 enum.py:1483(__and__)
        4    0.000    0.000    0.000    0.000 enum.py:669(__call__)
        3    0.000    0.000    0.000    0.000 _compiler.py:214(_compile_charset)
      3/2    0.000    0.000    0.000    0.000 _parser.py:172(getwidth)
        9    0.000    0.000    0.000    0.000 _parser.py:252(get)
        2    0.000    0.000    0.000    0.000 _parser.py:222(__init__)
       13    0.000    0.000    0.000    0.000 _parser.py:231(__next)
        9    0.000    0.000    0.000    0.000 _parser.py:162(__getitem__)
        3    0.000    0.000    0.000    0.000 _parser.py:307(_class_escape)
        2    0.000    0.000    0.000    0.000 _parser.py:444(_uniq)
       10    0.000    0.000    0.000    0.000 _parser.py:247(match)
        5    0.000    0.000    0.000    0.000 _parser.py:284(tell)
        2    0.000    0.000    0.000    0.000 _compiler.py:434(_get_literal_prefix)
        9    0.000    0.000    0.000    0.000 {method 'find' of 'bytearray' objects}
        2    0.000    0.000    0.000    0.000 _parser.py:73(__init__)
        4    0.000    0.000    0.000    0.000 enum.py:1074(__new__)
        4    0.000    0.000    0.000    0.000 _compiler.py:568(isstring)
        2    0.000    0.000    0.000    0.000 _compiler.py:465(_get_charset_prefix)
        8    0.000    0.000    0.000    0.000 {built-in method builtins.min}
        1    0.000    0.000    0.000    0.000 _compiler.py:396(_simple)
        5    0.000    0.000    0.000    0.000 _parser.py:158(__len__)
        2    0.000    0.000    0.000    0.000 _parser.py:954(fix_flags)
        3    0.000    0.000    0.000    0.000 _parser.py:109(__init__)
        2    0.000    0.000    0.000    0.000 {built-in method _sre.compile}
        4    0.000    0.000    0.000    0.000 _parser.py:79(groups)
        2    0.000    0.000    0.000    0.000 {built-in method fromkeys}
        6    0.000    0.000    0.000    0.000 {method 'get' of 'dict' objects}
        2    0.000    0.000    0.000    0.000 _parser.py:170(append)
        4    0.000    0.000    0.000    0.000 _compiler.py:426(_get_iscased)
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
        2    0.000    0.000    0.000    0.000 {method 'items' of 'dict' objects}
        1    0.000    0.000    0.000    0.000 {method 'insert' of 'list' objects}
        2    0.000    0.000    0.000    0.000 {built-in method builtins.ord}
        1    0.000    0.000    0.000    0.000 _parser.py:166(__setitem__)
```


After:

```
❯ ./.venv/bin/python -m calgen                                                          1m0s
         27613 function calls (27608 primitive calls) in 0.011 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.002    0.002    0.008    0.008 ical.py:105(calendar)
      389    0.001    0.000    0.007    0.000 ical.py:55(_event_to_ics)
      389    0.000    0.000    0.003    0.000 functional.py:224(wrapper)
     1556    0.003    0.000    0.003    0.000 {method 'strftime' of 'datetime.date' objects}
      389    0.000    0.000    0.002    0.000 text.py:397(slugify)
     1559    0.001    0.000    0.002    0.000 ical.py:16(_ical_escape)
      778    0.000    0.000    0.001    0.000 __init__.py:178(sub)
     1559    0.000    0.000    0.001    0.000 ical.py:31(_ical_fold)
       85    0.000    0.000    0.001    0.000 textwrap.py:373(wrap)
     9354    0.001    0.000    0.001    0.000 {method 'replace' of 'str' objects}
       85    0.000    0.000    0.001    0.000 textwrap.py:347(wrap)
      778    0.000    0.000    0.001    0.000 __init__.py:272(_compile)
      778    0.001    0.000    0.001    0.000 {method 'sub' of 're.Pattern' objects}
      389    0.001    0.000    0.001    0.000 <string>:2(__init__)
       85    0.000    0.000    0.000    0.000 textwrap.py:238(_wrap_chunks)
      389    0.000    0.000    0.000    0.000 {built-in method builtins.any}
       85    0.000    0.000    0.000    0.000 textwrap.py:341(_split_chunks)
        2    0.000    0.000    0.000    0.000 _compiler.py:738(compile)
      778    0.000    0.000    0.000    0.000 functional.py:226(<genexpr>)
       85    0.000    0.000    0.000    0.000 textwrap.py:157(_split)
     1194    0.000    0.000    0.000    0.000 {built-in method builtins.isinstance}
        2    0.000    0.000    0.000    0.000 _parser.py:970(parse)
       85    0.000    0.000    0.000    0.000 {method 'split' of 're.Pattern' objects}
1984/1981    0.000    0.000    0.000    0.000 {built-in method builtins.len}
        2    0.000    0.000    0.000    0.000 _parser.py:447(_parse_sub)
        2    0.000    0.000    0.000    0.000 _compiler.py:571(_code)
      778    0.000    0.000    0.000    0.000 {built-in method fromisoformat}
        2    0.000    0.000    0.000    0.000 _parser.py:507(_parse)
      389    0.000    0.000    0.000    0.000 {built-in method unicodedata.normalize}
       71    0.000    0.000    0.000    0.000 textwrap.py:197(_handle_long_word)
      765    0.000    0.000    0.000    0.000 {method 'append' of 'list' objects}
      389    0.000    0.000    0.000    0.000 {method 'encode' of 'str' objects}
        2    0.000    0.000    0.000    0.000 _compiler.py:509(_compile_info)
      389    0.000    0.000    0.000    0.000 {method 'decode' of 'bytes' objects}
       85    0.000    0.000    0.000    0.000 textwrap.py:112(__init__)
      389    0.000    0.000    0.000    0.000 {method 'lower' of 'str' objects}
      389    0.000    0.000    0.000    0.000 {method 'strip' of 'str' objects}
      389    0.000    0.000    0.000    0.000 {method 'values' of 'dict' objects}
      3/2    0.000    0.000    0.000    0.000 _compiler.py:37(_compile)
      241    0.000    0.000    0.000    0.000 {method 'join' of 'str' objects}
       85    0.000    0.000    0.000    0.000 textwrap.py:176(<listcomp>)
        3    0.000    0.000    0.000    0.000 _compiler.py:241(_optimize_charset)
       85    0.000    0.000    0.000    0.000 textwrap.py:143(_munge_whitespace)
        4    0.000    0.000    0.000    0.000 enum.py:1483(__and__)
       85    0.000    0.000    0.000    0.000 {method 'pop' of 'list' objects}
       71    0.000    0.000    0.000    0.000 {built-in method builtins.sum}
        3    0.000    0.000    0.000    0.000 _compiler.py:214(_compile_charset)
        9    0.000    0.000    0.000    0.000 _parser.py:252(get)
        4    0.000    0.000    0.000    0.000 enum.py:669(__call__)
      3/2    0.000    0.000    0.000    0.000 _parser.py:172(getwidth)
       13    0.000    0.000    0.000    0.000 _parser.py:231(__next)
       85    0.000    0.000    0.000    0.000 {method 'reverse' of 'list' objects}
        2    0.000    0.000    0.000    0.000 _parser.py:222(__init__)
        9    0.000    0.000    0.000    0.000 _parser.py:162(__getitem__)
        3    0.000    0.000    0.000    0.000 _parser.py:307(_class_escape)
        2    0.000    0.000    0.000    0.000 _parser.py:444(_uniq)
       10    0.000    0.000    0.000    0.000 _parser.py:247(match)
        5    0.000    0.000    0.000    0.000 _parser.py:284(tell)
        2    0.000    0.000    0.000    0.000 _compiler.py:434(_get_literal_prefix)
        1    0.000    0.000    0.000    0.000 _compiler.py:396(_simple)
        4    0.000    0.000    0.000    0.000 _compiler.py:568(isstring)
        2    0.000    0.000    0.000    0.000 _compiler.py:465(_get_charset_prefix)
        9    0.000    0.000    0.000    0.000 {method 'find' of 'bytearray' objects}
        8    0.000    0.000    0.000    0.000 {built-in method builtins.min}
        5    0.000    0.000    0.000    0.000 _parser.py:158(__len__)
        2    0.000    0.000    0.000    0.000 _parser.py:954(fix_flags)
        4    0.000    0.000    0.000    0.000 _parser.py:79(groups)
        4    0.000    0.000    0.000    0.000 enum.py:1074(__new__)
        6    0.000    0.000    0.000    0.000 {method 'get' of 'dict' objects}
        2    0.000    0.000    0.000    0.000 _parser.py:73(__init__)
        3    0.000    0.000    0.000    0.000 _parser.py:109(__init__)
        2    0.000    0.000    0.000    0.000 {built-in method fromkeys}
        2    0.000    0.000    0.000    0.000 {built-in method _sre.compile}
        2    0.000    0.000    0.000    0.000 _parser.py:170(append)
        4    0.000    0.000    0.000    0.000 _compiler.py:426(_get_iscased)
        2    0.000    0.000    0.000    0.000 {built-in method builtins.ord}
        1    0.000    0.000    0.000    0.000 _parser.py:166(__setitem__)
        2    0.000    0.000    0.000    0.000 {method 'items' of 'dict' objects}
        1    0.000    0.000    0.000    0.000 {method 'insert' of 'list' objects}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```